refactor: 도메인 로직 recoil 파일에서 분리
